add cronjob for tempban system add cronjob for autoChannelDelete system fix userinfo if user == null some improvements report system

+++ src/events/autoDelete/autoDeleteChannel.js
@@ -1,6 +1,7 @@
+const schedule = require("node-schedule");
 var client;
 
-async function userTempBanCheck() {
+async function autoDeleteChannelCheck() {
     const autoDeleteChannelIds = ["1049671515836264499"];
 
     if (!autoDeleteChannelIds) {
@@ -13,14 +14,29 @@ async function userTempBanCheck() {
 }
 
 async function autoDeleteMessages(autoDeleteChannelId) {
-  const autoDeleteChannel = await client.channels.fetch(autoDeleteChannelId);
+  return new Promise(async (resolve) => {
+  
+    let autoDeleteChannel = ``;
+
+    try {
+      autoDeleteChannel = await client.channels.fetch(autoDeleteChannelId);
+    } catch (error) {}
+
+  if (!autoDeleteChannel) {
+    return resolve(null);
+  }
 
   await autoDeleteChannel.bulkDelete(100, true)
+  return resolve(null);
+  })
 }
 
 async function init(_client) {
   client = _client;
-  setInterval(userTempBanCheck, 86400000);
+  const cronJobTempBan = schedule.scheduleJob({hour: 23, minute: 00}, function() {
+    console.log(`\x1b[32mCRONJOB | Auto-Delete-Channel ausgeführt\x1b[0m`);
+    autoDeleteChannelCheck()
+  });
 }
 
 module.exports.init = init;
